pyproject.toml
[project]
name = "skill-framework"
version = "0.1.0"
description = "Skills meta-tool architecture for Google ADK agents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=0.1.0",
    "pyyaml>=6.0",
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.42.30",
    "openinference-instrumentation-google-adk>=0.1.0",
    "arize-phoenix-otel>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
web = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/skill_framework"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.13",
    "types-pyyaml>=6.0.12.20250915",
]
